Implement configurable remote host in GD1 backend
This commit implements all the bits which are required to make `gluster`
connect to a remote host. The config variable `gd1-remote-host` allows
the user to specify the hostname or IP address of the GD1 instance to
which the exporter should connect.

If the variable is missing from the config file, the value of
`GConfig.GlusterRemoteHost` is set to `localhost`.