Added TTL to all Metrics exported
Reseting all metrics causes missing entries since the Prometheus
server may pull the metrics from exporter when it was reset.
To avoid the problem, TTL option is provided(Default is 2 minutes).

If a Metric is not exported for the TTL period, then it deletes
that metric from exporter to avoid stale entries.

Fixes: #158
Signed-off-by: Aravinda VK <avishwan@redhat.com>